Updated the api

+++ WebAPI/Heritage/Heritage/Models/PlaceMarker.cs
@@ -11,7 +11,7 @@ namespace Heritage.Models
         public long PlaceMarkerId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public long PlaceMarkerTypeID { get; set; }
+        public int PlaceMarkerTypeID { get; set; }
         public string Latitude { get; set; }
         public string Longitude { get; set; }
     }
